Remove @project tag from BigNumber.php
Hi guys,

When using Doctrine packages in our project, we get this conflict with your use of the "@project" tag in BigNumber.php:

"Doctrine\\Common\\Annotations\\AnnotationException: [Semantical Error] The annotation \"@project\" in class Aza\\Components\\Math\\BigNumber was never imported. Did you maybe forget to add a \"use\" statement for this annotation?"

It works if we remove the @project tag. 

Thanks!